Add files via upload
This Blackjack game developed using basic level of skill in python using lists,functions,loops,modules,operators and etc


Rules In Blackjack:

Everyone plays against the dealer. Players receive all cards face up, and the dealer’s first card is face up and the second is face down. The object of the game is to get closer to 21 than the dealer without going over 21. If a hand goes over 21, it is called a “bust” or “break,” and the wager is lost. In 21, Jacks, Queens, Kings, and 10s count as 10. An Ace may be played as a one or an 11. All other cards are played at face value.

When you receive your first two cards, you may either “stand” or “hit.” When you “stand,” it means you feel you are close enough to 21 and no longer wish any additional cards, indicated by waving off with your hand. If you wish to receive another card or “hit,” tap or scratch the table behind your wager with your finger.  In either situation, you will never touch the cards; everything is communicated using hand signals. You may draw as many cards as you want until you are close to 21 or until you “bust.”

If you are closer to 21 than the dealer, you win and are paid an amount equal to your original wager. If your hand is less than the dealer’s, you lose. If the dealer’s hand “busts” or “breaks,” you win as well. Ties are a standoff or “push,” and your bet remains on the table.

If your initial two cards total 21, any Ace with a 10, Jack, Queen, or King, you have a Blackjack. Blackjack is paid either 6 to 5 or 3 to 2, depending on the type of Blackjack you are playing.


Try and give me feedback on thiss development...